refactor: update request and response types, remove deprecated DTOs, and enhance middleware structure
