--- server/src/types/dto/requests/auth.requests.ts
@@ -1,24 +1,0 @@
-/**
- * User registration request DTO
- */
-export interface RegisterUserRequest {
-  username: string;
-  password: string;
-  firstName?: string;
-  lastName?: string;
-}
-
-/**
- * User login request DTO
- */
-export interface LoginRequest {
-  username: string;
-  password: string;
-}
-
-/**
- * Token refresh request DTO
- */
-export interface RefreshTokenRequest {
-  refreshToken: string;
-}
