--- server/src/types/express/index.d.ts
@@ -1,9 +1,0 @@
-import { AuthenticatedUser } from "types/domain/user";
-
-declare global {
-  namespace Express {
-    interface Request {
-      authenticatedUser?: AuthenticatedUser;
-    }
-  }
-}
